refactor: convert flow view route for vite
