+++ pages/FlowViewPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import FlowVisualizer from '../../../components/FlowVisualizer.jsx';
+import FlowVisualizer from '../components/FlowVisualizer.jsx';
 
 export default function FlowViewPage({ flowId }) {
   return (
